Per our discussion last night, moving TLSHostingPower to network.server and beginning the 'Kirby' pattern.
